pyproject.toml
[tool.poetry]
name = "to"
version = "0.1.0"
description = "Bookmark search API"
authors = ["Tom Bulled <26026015+tombulled@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
PyYAML = "^6.0"
thefuzz = "^0.19.0"
furl = "^2.1.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
mypy = "^0.961"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
